AI/requirements.txt
langchain==0.1.13
langchain-community==0.0.33
langchain-openai==0.0.8
openai==1.14.3
tiktoken==0.5.2
faiss-cpu==1.7.4
jq
python-dotenv==1.0.1
tqdm
